apps/views/chat.py
import logging
from rest_framework import viewsets
from rest_framework.response import Response
from rest_framework import status
# from django_filters.rest_framework import DjangoFilterBackend
from rest_framework.viewsets import ModelViewSet

from apps.utils import CustomAuthenticated

log = logging.getLogger(__file__)
